Separate phonon and phonon_qt5
Having dual-purpose Nix expressions for phonon and each of its backends
was more trouble than simply having one set of expressions for Qt4 and
one for Qt5. On top of that, the only version of
phonon-backend-gstreamer that supports Qt5 requires gstreamer-1.0, but
the Qt4 build will always use gstreamer-0.10 at runtime (regardless of
how you compile it). The whole thing is a mess. It's better to think of
the Qt4 and Qt5 versions as completely unrelated packages that happen to
have overlapping names.